+++ client/heatmap_script.js
@@ -1,4 +1,36 @@
+import request from 'request';
 
+const accessToken = 'BQACX0bVaTJ99cc7DJwDua1RqDnpj-VXOPacGoI51GypyRWNwCS2wJdlkd--TIFoVXpSNCppN1D_Xv4RvJhu8d9gHzwVOVyuzgdpzaAw96yjWAOTtXtvnQDcj8TrHXq3aIKxbgPSbIdnJJDz74jcb0lNbQztUxoemvnvdyZ0DtwyzO8QAuNXlNiMgfUNaFFm7XU';
+  
+const countries = [];
+
+// Set the headers
+const headers = {
+  'Authorization': 'Bearer ' + accessToken
+}
+
+// Configure the request
+const options = {
+  url: 'https://api.spotify.com/v1/browse/categories',
+  method: 'GET',
+  headers: headers
+}
+
+// Start the request
+request(options, function (error, response, body) {
+  if (!error && response.statusCode == 200) {
+    // Parse the response body
+    const data = JSON.parse(body);
+
+    // Loop through the genres and add each country to the array
+    data.categories.items.forEach(genre => {
+      countries.push(genre.country);
+    });
+
+    // Print the array of countries
+    console.log(countries);
+  }
+});
